P{ut names pkg in strins
