+++ starport/pkg/cosmosgen/generate_javascript.go
@@ -11,11 +11,11 @@ import (
 	"github.com/mattn/go-zglob"
 	"github.com/tendermint/starport/starport/pkg/cosmosanalysis/module"
 	"github.com/tendermint/starport/starport/pkg/gomodule"
-	"github.com/tendermint/starport/starport/pkg/names"
 	"github.com/tendermint/starport/starport/pkg/nodetime/sta"
 	tsproto "github.com/tendermint/starport/starport/pkg/nodetime/ts-proto"
 	"github.com/tendermint/starport/starport/pkg/nodetime/tsc"
 	"github.com/tendermint/starport/starport/pkg/protoc"
+	"github.com/tendermint/starport/starport/pkg/xstrings"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -226,7 +226,7 @@ func (g *jsGenerator) generateVuexModuleLoader() error {
 
 		var (
 			fullPath = filepath.Dir(pathrel)
-			fullName = names.FormatUsername(strcase.ToCamel(strings.ReplaceAll(fullPath, "/", "_")))
+			fullName = xstrings.FormatUsername(strcase.ToCamel(strings.ReplaceAll(fullPath, "/", "_")))
 			path     = filepath.Base(fullPath)
 			name     = strcase.ToCamel(path)
 		)
