--- starport/pkg/names/names.go
@@ -1,26 +1,0 @@
-package names
-
-import (
-	"strings"
-	"unicode"
-)
-
-// FormatUsername formats a username to make it usable as a variable
-func FormatUsername(s string) string {
-	return NoDash(NoNumberPrefix(s))
-}
-
-// NoDash removes dash from the string
-func NoDash(s string) string {
-	return strings.ReplaceAll(s, "-", "")
-}
-
-// NoNumberPrefix adds a underscore at the beginning of the string if it stars with a number
-// this is used for package of proto files template because the package name can't start with a string
-func NoNumberPrefix(s string) string {
-	// Check if it starts with a digit
-	if unicode.IsDigit(rune(s[0])) {
-		return "_" + s
-	}
-	return s
-}
