+++ starport/pkg/xstrings/xstrings.go
@@ -1,5 +1,10 @@
 package xstrings
 
+import (
+	"strings"
+	"unicode"
+)
+
 // AllOrSomeFilter filters elems out from the list as they  present in filterList and
 // returns the remaning ones.
 // if filterList is empty, all elems from list returned.
@@ -40,4 +45,24 @@ func List(n int, do func(i int) string) []string {
 	}
 
 	return list
+}
+
+// FormatUsername formats a username to make it usable as a variable
+func FormatUsername(s string) string {
+	return NoDash(NoNumberPrefix(s))
+}
+
+// NoDash removes dash from the string
+func NoDash(s string) string {
+	return strings.ReplaceAll(s, "-", "")
+}
+
+// NoNumberPrefix adds a underscore at the beginning of the string if it stars with a number
+// this is used for package of proto files template because the package name can't start with a string
+func NoNumberPrefix(s string) string {
+	// Check if it starts with a digit
+	if unicode.IsDigit(rune(s[0])) {
+		return "_" + s
+	}
+	return s
 }
