+++ starport/templates/module/create/new_ibc.go
@@ -4,14 +4,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/tendermint/starport/starport/pkg/names"
-
-	"github.com/gobuffalo/plush"
-
-	"github.com/tendermint/starport/starport/templates/module"
-
 	"github.com/gobuffalo/genny"
+	"github.com/gobuffalo/plush"
 	"github.com/gobuffalo/plushgen"
+	"github.com/tendermint/starport/starport/pkg/xstrings"
+	"github.com/tendermint/starport/starport/templates/module"
 )
 
 // New ...
@@ -39,7 +36,7 @@ func NewIBC(opts *CreateOptions) (*genny.Generator, error) {
 	ctx.Set("title", strings.Title)
 
 	// Used for proto package name
-	ctx.Set("formatOwnerName", names.FormatUsername)
+	ctx.Set("formatOwnerName", xstrings.FormatUsername)
 
 	g.Transformer(plushgen.Transformer(ctx))
 	g.Transformer(genny.Replace("{{moduleName}}", opts.ModuleName))
